+++ docs/changelog.rst
@@ -67,6 +67,7 @@ Changes in dev
   way the user has specified them
 - Fixed various errors
 - Added many tests
+- Added pre and post bulk-insert signals
 
 Changes in v0.4
 ===============